feat: add Docker container for frontend
Also, move and modify the `dev-database` `docker-compose.yml` to run the
frontend in addition to the database. This means you can do a simple
`docker-compose up` in the project root directory to run a local
development system.

NB! Database tables other than `User` is not yet supported. This means
the frontend *will* crash if it tries interact with the other tables,
such as `UserMembership`.